seperate JD send_request
